src/requirements/development.txt
-r ./base.txt
pytest
pytest-django
pytest-cov
black
faker
django-debug-toolbar
coverage
django-coverage-plugin